requirements.txt
fastapi>=0.111.0,<0.113.0
fastapi-debug-toolbar==0.6.3
pydantic>=2.7.0,<3.0.0
pydantic_extra_types==2.8.2
uvicorn==0.30.1
pymongo==4.8.0
motor==3.5.0
pydantic-mongo==2.3.0
